+++ NEWS.md
@@ -8,7 +8,8 @@
 * Major changes:
   * Function `read_airfares()` is working again. Closes [#30](https://github.com/ipeaGIT/flightsbr/issues/30)
   * Function `read_flights()` with fixed decimal values in numeric columns. Closes [#43](https://github.com/ipeaGIT/flightsbr/issues/43)
-  
+  * Function `read_airports()` with fixed values for `"altitude"` column. Closes [#42](https://github.com/ipeaGIT/flightsbr/issues/42)
+    
 * Minor changes:
   * Internally check of the consistency of date inputs. The date input must be consistent in either a 6-digit format `yyyymm` OR a 4-digit format `yyyy`.
   * New support function `latest_airfares_date()`